% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Summary.R
\name{summary.eefAnalytics}
\alias{summary.eefAnalytics}
\title{Summary for a fitted model represented by an \code{eefAnalytics} object.}
\usage{
\method{summary}{eefAnalytics}(object, ...)
}
\arguments{
\item{object}{Object of class \code{eefAnalytics}}

\item{...}{Additional arguments of \code{\link[base]{summary}}}
}
\value{
Returns relevant summary including Rhat and effective sample sizes.
}
\description{
Summary for a fitted model represented by an \code{eefAnalytics} object.
}
